pyproject.toml
[tool.poetry]
name = "autify"
version = "0.1.0"
description = ""
authors = ["Matheus Mello de Lima <mellomatheuslima@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.30.0"
pydantic = "^1.10.7"
pendulum = "^2.1.2"
spotipy = "^2.23.0"
python-dotenv = "^1.0.0"
fastapi = "^0.95.1"

[tool.poetry.scripts]
app = "src.app:run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
